requirements.txt
aiohttp==3.7.3
aioresponses==0.7.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
multidict==5.1.0
packaging==20.8
paho-mqtt==1.5.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.0
pytest-asyncio==0.14.0
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
